+++ src/meteor.js
@@ -8,7 +8,8 @@ const build = (env) => (
       return reject("Please provide a server as the second argument");
     }
 
-    let buildAction = `meteor build .build --architecture os.linux.x86_64 --server ${meteorServer}`;
+    let buildAction = `meteor build ${env.METEOR_OUTPUT_DIR}`;
+    buildAction += ` --architecture os.linux.x86_64 --server ${meteorServer}`;
     if (process.argv[2]) {
       buildAction += ` --mobile-settings ${process.argv[2]}`;
     }