+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Support for custom build directory with `METEOR_OUTPUT_DIR`. (#81)
+
 ## 0.2.7
 
 * Standalone action for preparing Android SDKs